Added custom depth limiting and more AI tests
